Real fix for #113
Apparently the cause of this issue is that Tomcat checks the System
classloader before it checks any others.  Unfortunately, the sbt-
lauch.jar is on the system classloader.  It appears to contain a version
of the Scala standard lib that has been run through Proguard.  As a
result, it is missing a bunch of methods.  To fix this I modified the
Tomcat classloader so it uses the root classloader instead of the system
classloader.  This is kind of a hack because Tomcat doesn't provided a
way to do this.