Use the webapp path provided to containerLaunchCmd
This updates the Payara Micro test to use the webapp path provided as an
argument to `containerLaunchCmd`.  The value of `path` is an absolute
path to either *target/webapp* or *target/webapp-quick*, depending on
whether `container:test` or `container:quicktest` is used.

This is largely a noop as `container:quicktest` fails, possibly because
Payara Micro uses a separate classloader for the webapp than for the
forked JVM.