Create a webapp components plugin
To save on unnecessary IO, we want to be able to identify all of the
components of a webapp without needlessly copying files around.

A webapp consists of three sets of files:

* resources (HTML/CSS/JS files, images, deployment descriptor, etc.)
* .class files to be loaded by the container classloader
* .jar files to be loaded by the container classloader

In most cases, resources are static and kept in the `src/main/webapp`
directory.  Occasionally they might be generated (e.g. via a minifier, a
JS compiler, etc.).

The webapp components plugin allows the user to specify/override the
locations of these files.

The webapp components plugin will be used by a downstream container
runner to launch a container without needing to copy static files
around, prepare an exploded .war file, or package a .war file.