Extract the container main class names to settings (#978)
This creates two new hidden settings:

* `WarPackageRunnerPlugin/warMainClass`
* `WebappComponentsRunnerPlugin/webappComponentsRunnerMainClass`

These are used when forking new container processes to specify the main
method of the forked JVM.

Note that for WarPackageRunnerPlugin, we dropped the `-jar` argument and
converted to `-cp` so that we can specify the main class manually.  This
also lets us use multiple libraries in the `War` configuration, if ever
needed.